2019/8-azure-service-bus-messaging/requirements.txt
flask
azure-servicebus==7.0.0b5